# openapi.yaml
openapi: 3.1.3
info:
  title: Ludus Extension API
  version: "1.0"
  description: API for managing scenarios, data, and topology in Ludus CTFd extension.
servers:
  - url: http://127.0.0.1:5000
components:
  securitySchemes:
    ApiKeyAuth: 
      type: apiKey 
      in: header
      name: X-API-Key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
      required:
        - error
security:
  - ApiKeyAuth: []
paths:
...
  /range/access:
    get:
      summary: Get WireGuard access configs for pool users
      description: Download WireGuard configuration files for all users in a pool as a base64-encoded ZIP file
      tags:
        - Ludus Range Sharing
      parameters:
        - name: poolId
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9]{6}$"
          description: Pool ID to get users from
          example: "ABC123"
      responses:
        '200':
          description: Base64-encoded ZIP file containing WireGuard configurations
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                    description: Suggested filename for the ZIP file
                    example: "ABC123.zip"
                  data:
                    type: string
                    format: byte
                    description: Base64-encoded ZIP file containing WireGuard configs in wireguard-configs/ folder
                    example: "UEsDBBQAAAAIAG5kJVnzNwAA8T..."
                  size:
                    type: integer
                    description: Size of the ZIP file in bytes
                    example: 2048
                required:
                  - filename
                  - data
                  - size
              examples:
                successful_download:
                  summary: Successful ZIP file generation
                  value:
                    filename: "ABC123.zip"
                    data: "UEsDBBQAAAAIAG5kJVnzNwAA8TEAAABAAAAA..."
                    size: 2048
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Pool or configs not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_configs:
                  summary: No WireGuard configs found
                  value:
                    error: "No configs found"
                pool_not_found:
                  summary: Pool not found
                  value:
                    error: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                zip_creation_failed:
                  summary: ZIP file creation failed
                  value:
                    error: "Failed to create ZIP file"
...
  /stats/proxmox:
    get:
      summary: Get Proxmox cluster statistics
      description: Retrieves comprehensive statistics about the Proxmox cluster including users, templates, VMs, topologies, scenarios, roles, pools, and node information
      operationId: getProxmoxStatistics
      tags: 
        - Proxmox
      responses:
        '200':
          description: Successfully retrieved Proxmox statistics
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                  - templates
                  - vms
                  - numberOfTopologies
                  - numberOfScenarios
                  - numberOfRoles
                  - numberOfPools
                  - cpuUsagePercentage
                  - maxCpu
                  - memoryUsedGiB
                  - memoryTotalGiB
                  - memoryFreeGiB
                  - diskUsedGiB
                  - diskTotalGiB
                  - uptimeFormatted
                properties:
                  users:
                    type: integer
                    description: Total number of users (pools) in the Proxmox cluster (excluding ADMIN and SHARED)
                    minimum: 0
                    example: 5
                  templates:
                    type: integer
                    description: Total number of VM templates available
                    minimum: 0
                    example: 6
                  vms:
                    type: integer
                    description: Total number of virtual machines (excluding templates)
                    minimum: 0
                    example: 4
                  numberOfTopologies:
                    type: integer
                    description: Total number of topology configurations available
                    minimum: 0
                    example: 12
                  numberOfScenarios:
                    type: integer
                    description: Total number of CTFd scenarios available
                    minimum: 0
                    example: 8
                  numberOfRoles:
                    type: integer
                    description: Total number of Ansible roles available in Ludus
                    minimum: 0
                    example: 25
                  numberOfPools:
                    type: integer
                    description: Total number of pools in the system
                    minimum: 0
                    example: 7
                  cpuUsagePercentage:
                    type: number
                    format: double
                    description: Current CPU usage as a percentage
                    minimum: 0
                    maximum: 100
                    example: 0.17
                  maxCpu:
                    type: integer
                    description: Maximum number of CPU cores available
                    minimum: 1
                    example: 16
                  memoryUsedGiB:
                    type: number
                    format: double
                    description: Currently used memory in GiB (Gibibytes)
                    minimum: 0
                    example: 6.24
                  memoryTotalGiB:
                    type: number
                    format: double
                    description: Total memory capacity in GiB (Gibibytes)
                    minimum: 0
                    example: 62.76
                  memoryFreeGiB:
                    type: number
                    format: double
                    description: Available memory in GiB (Gibibytes)
                    minimum: 0
                    example: 56.53
                  diskUsedGiB:
                    type: number
                    format: double
                    description: Currently used disk space in GiB (Gibibytes)
                    minimum: 0
                    example: 66.51
                  diskTotalGiB:
                    type: number
                    format: double
                    description: Total disk capacity in GiB (Gibibytes)
                    minimum: 0
                    example: 822.43
                  uptimeFormatted:
                    type: string
                    description: Human-readable uptime format (e.g., "2w 4d 3h")
                    pattern: '^(\d+w\s)?(\d+d\s)?(\d+h)?$'
                    example: "2w 4d 3h"
              example:
                users: 5
                templates: 6
                vms: 4
                numberOfTopologies: 12
                numberOfScenarios: 8
                numberOfRoles: 25
                numberOfPools: 7
                cpuUsagePercentage: 0.17
                maxCpu: 16
                memoryUsedGiB: 6.24
                memoryTotalGiB: 62.76
                memoryFreeGiB: 56.53
                diskUsedGiB: 66.51
                diskTotalGiB: 822.43
                uptimeFormatted: "2w 4d 3h"
        '401':
          description: Unauthorized - Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error - Proxmox configuration missing or connection failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...